reorganize the file structure, add more scripts
