+++ openwhisk/actions/actions_delete.sh
@@ -1,0 +1,7 @@
+wsk -i action delete updateDriver
+
+wsk -i action delete searchPassenger
+
+wsk -i action delete updatePassenger
+
+wsk -i action delete searchDriver
