+++ openwhisk/docker/Dockerfile_mongo
