+++ openwhisk/docker/Dockerfile_redis
