+++ openwhisk/scripts/deploy_redis.sh
@@ -1,0 +1,3 @@
+kubectl apply -f redis-deployment.yaml
+
+kubectl apply -f redis-node-port.yaml
