+++ openwhisk/scripts/mongodb-cluster-ip-service.yaml
@@ -1,0 +1,11 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: mongodb-cluster-ip-service
+spec:
+  type: ClusterIP
+  selector:
+    component:  redis 
+  ports:
+    - port: 27017 
+      targetPort: 27017
