+++ openwhisk/scripts/redis-cluster-ip-service.yaml
@@ -1,0 +1,11 @@
+apiVersion: v1
+kind: Service
+metadata:
+  name: redis-cluster-ip-service
+spec:
+  type: ClusterIP
+  selector:
+    component:  redis 
+  ports:
+    - port: 6379 
+      targetPort: 6379
