+++ openwhisk/scripts/start_OW.sh
@@ -1,0 +1,3 @@
+minikube start
+
+minikube ssh -- sudo ip link set docker0 promisc on
